downloader_service/requirements.txt
asgiref==3.10.0
boto3==1.40.58
botocore==1.40.58
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-storages==1.14.6
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
jmespath==1.0.1
packaging==25.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
yt-dlp==2025.10.22
